web service MDL-12886 implement authentication token systemprototype

+++ user/external.php
@@ -86,15 +86,15 @@ final class user_external extends moodle_external {
      * @return boolean true if success
      */
     static function tmp_delete_user($params) {
-        global $DB;
+        global $DB,$USER;
+
         $user = $DB->get_record('user', array('username'=>$params['username'], 'mnethostid'=>$params['mnethostid']));
-    /// PLEASE UNCOMMENT HERE ONCE AUTHENTICATION IS IMPLEMENTED - $USER/context need to be set here
-//        if (has_capability('moodle/user:delete', get_context_instance(CONTEXT_SYSTEM))) {
+        if (has_capability('moodle/user:delete', get_context_instance(CONTEXT_SYSTEM))) {
             return delete_user($user); //this function is in moodlelib.php
-//        }
-//        else {
-//            throw new moodle_exception('couldnotdeleteuser');
-//        }
+        }
+        else {
+            throw new moodle_exception('couldnotdeleteuser');
+        }
     }
